fix: Don't draw areas on init if layer is marked as hidden

+++ src/app/map/[id]/components/Map.tsx
@@ -107,7 +107,7 @@ export default function Map({
 
   // draw existing turfs
   useEffect(() => {
-    if (!turfs || !draw) return;
+    if (!turfs || !draw || !viewConfig?.showTurf) return;
 
     draw.deleteAll();
 
@@ -119,7 +119,7 @@ export default function Map({
         geometry: turf.polygon,
       });
     });
-  }, [turfs, draw]);
+  }, [turfs, draw, viewConfig?.showTurf]);
 
   // Hover behavior
   useEffect(() => {
